Release App Phoenix v1.0.0 (Official Release)

+++ controllers/task.js
@@ -3,18 +3,29 @@ const asyncWrapper = require("../middleware/async");
 const { createCustomerError } = require("../error/custom-error");
 
 const getAllItems = asyncWrapper(async (req, res) => {
-  const tasks = await Task.find(req.query);
+  const filter = { createdBy: req.user.userId };
+
+  if (req.query.completed === "true") {
+    filter.completed = true;
+  } else if (req.query.completed === "false") {
+    filter.completed = false;
+  }
+  const tasks = await Task.find(filter).sort("createdAt");
   res.status(200).json({ tasks });
 });
 
 const createTask = asyncWrapper(async (req, res) => {
+  req.body.createdBy = req.user.userId;
   const task = await Task.create(req.body);
   res.status(201).json({ task });
 });
 
 const getTask = asyncWrapper(async (req, res) => {
-  const { id: taskID } = req.params;
-  const task = await Task.findOne({ _id: taskID });
+  const {
+    user: { userId },
+    params: { id: taskID },
+  } = req;
+  const task = await Task.findOne({ _id: taskID, createdBy: userId });
   if (!task) {
     return next(createCustomerError("There is no task with the id: ${taskID}"));
   }
@@ -22,20 +33,30 @@ const getTask = asyncWrapper(async (req, res) => {
 });
 
 const deleteTask = asyncWrapper(async (req, res) => {
-  const { id: taskID } = req.params;
-  const task = await Task.findOneAndDelete({ _id: taskID });
+  const {
+    user: { userId },
+    params: { id: taskID },
+  } = req;
+  const task = await Task.findOneAndDelete({ _id: taskID, createdBy: userId });
   if (!task) {
-    return next(createCustomerError("There is no task with the id: ${taskID}"));
+    return next(createCustomerError(`There is no task with the id: ${taskID}`));
   }
   res.status(200).json({ task });
 });
 
 const updateTask = asyncWrapper(async (req, res) => {
-  const { id: taskID } = req.params;
-  const task = await Task.findByIdAndUpdate({ _id: taskID }, req.body, {
-    new: true,
-    runValidators: true,
-  });
+  const {
+    user: { userId },
+    params: { id: taskID },
+  } = req;
+  const task = await Task.findByIdAndUpdate(
+    { _id: taskID, createdBy: userId },
+    req.body,
+    {
+      new: true,
+      runValidators: true,
+    }
+  );
   if (!task) {
     return next(createCustomerError("There is no task with the id: ${taskID}"));
   }
